requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
sentence-transformers
faiss-cpu
pydantic
python-dotenv
requests
beautifulsoup4
PyMuPDF

#KT 믿음 쓰기 위해 필요한 모델
#torch --index-url https://download.pytorch.org/whl/cpu
#transformers
#accelerate

# PostgreSQL + pgvector 지원
psycopg2-binary
numpy
PyPDF2